Finalizado os CRUDs conforme exercicio proposto

+++ CatalogoFilmes/Controllers/FilmeController.cs
@@ -38,19 +38,41 @@ namespace CatalogoFilmes.Controllers
         }
 
 
+        [HttpGet("ObterPorNome")]
+        public ActionResult<Filme> ObterPorNome([FromQuery] string nome)
+        {
+            var filme = _repository.ObterPorNome(nome);
+            if (filme is null)
+            {
+                return NotFound("Filme nao encontrado");
+            }
+            return Ok(filme);
+        }
+
+
         [HttpPost]
         public ActionResult Post([FromBody] Filme filme)
         {
             _repository.Adicionar(filme);
-            return Ok();
+            //return Ok();
+            return Ok("Codigo do filme INCLUIDO: " + filme.Id);
+
         }
 
 
         [HttpPut("{id:int}")]
         public ActionResult Put(int id, [FromBody] Filme filme)
         {
-            _repository.Atualizar(filme);
-            return NoContent();
+            var filmeAtualizar = _repository.ObterPorId(id);
+            if (filmeAtualizar is null)
+            {
+                return BadRequest("Filme nao encontrado");
+            }
+            else
+            {
+                _repository.Atualizar(filme);
+                return Ok("Codigo do filme ATUALIZADO: " + filme.Id);
+            }
         }
 
 
@@ -58,9 +80,16 @@ namespace CatalogoFilmes.Controllers
         [HttpDelete("{id:int}")]
         public ActionResult Delete(int id)
         {
-            if 
-            _repository.Remover(id);
-            return NoContent();
+            var filme = _repository.ObterPorId(id);
+            if (filme is null)
+            {
+                return BadRequest("Filme nao encontrado");
+            }
+            else
+            {
+                _repository.Remover(id);
+                return Ok("Codigo do filme DELETADO: " + id);
+            }
         }
     }
 }
